Adding updateInterval() to LinkCable, LinkWireless, and LinkUniversal

+++ lib/LinkMobile.hpp
@@ -707,7 +707,7 @@ class LinkMobile {
 
   struct Config {
     u32 timeout;
-    u32 timerId;
+    u8 timerId;
   };
 
   /**
